add architecture docs
