+++ src/Features/Sensors/EcoData.Sensors.Api/SensorsApiExtensions.cs
@@ -1,3 +1,4 @@
+using EcoData.Sensors.Api.Endpoints;
 using Microsoft.AspNetCore.Routing;
 
 namespace EcoData.Sensors.Api;
@@ -7,9 +8,10 @@ public static class SensorsApiExtensions
     public static IEndpointRouteBuilder MapSensorsApiEndpoints(this IEndpointRouteBuilder app)
     {
         app.MapSensorEndpoints();
+        app.MapSensorReadingEndpoints();
         app.MapSensorHealthEndpoints();
+        app.MapSensorHealthConfigEndpoints();
         app.MapReferenceDataEndpoints();
-        app.MapPushEndpoints();
 
         return app;
     }
