+++ src/Features/Sensors/EcoData.Sensors.Contracts/Dtos/HeartbeatResponse.cs
@@ -1,0 +1,3 @@
+namespace EcoData.Sensors.Contracts.Dtos;
+
+public sealed record HeartbeatResponse(string Message, Guid SensorId, DateTimeOffset Timestamp);
